fixing detekt
