+++ .idea/misc.xml
@@ -10,6 +10,7 @@
         <entry key="..\:/Users/paliy/StudioProjects/Android-Universal-Recycler-View-Adapter/app/src/main/res/layout/item_heading.xml" value="0.25" />
         <entry key="..\:/Users/paliy/StudioProjects/Android-Universal-Recycler-View-Adapter/app/src/main/res/layout/item_list.xml" value="0.1" />
         <entry key="..\:/Users/paliy/StudioProjects/Android-Universal-Recycler-View-Adapter/app/src/main/res/layout/item_user.xml" value="0.2" />
+        <entry key="..\:/Users/paliy/StudioProjects/Android-Universal-Recycler-View-Adapter/universal_adapter/src/main/res/layout/item_simple.xml" value="0.1" />
       </map>
     </option>
   </component>
