+++ app/build.gradle
@@ -2,6 +2,8 @@ apply plugin: 'com.android.application'
 
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-kapt'
+apply plugin: "io.gitlab.arturbosch.detekt"
+
 
 android {
     compileSdkVersion 29
