+++ app/src/main/java/com/techpaliyal/androidkotlinmvvm/ui/multiple_view/HeadingModel.kt
@@ -1,7 +1,7 @@
 package com.techpaliyal.androidkotlinmvvm.ui.multiple_view
 
 import com.techpaliyal.androidkotlinmvvm.R
-import com.yogeshpaliyal.universal_adapter.listener.UniversalViewType
+import com.yogeshpaliyal.universalAdapter.listener.UniversalViewType
 
 data class HeadingModel(val title: String) : UniversalViewType,SchoolListing {
