+++ app/src/main/java/com/techpaliyal/androidkotlinmvvm/ui/multiple_view/ListItemModel.kt
@@ -1,7 +1,7 @@
 package com.techpaliyal.androidkotlinmvvm.ui.multiple_view
 
 import com.techpaliyal.androidkotlinmvvm.R
-import com.yogeshpaliyal.universal_adapter.listener.UniversalViewType
+import com.yogeshpaliyal.universalAdapter.listener.UniversalViewType
 
 data class ListItemModel(val name: String) : UniversalViewType,SchoolListing {
     override fun getLayoutId(): Int {
