+++ app/src/main/java/com/techpaliyal/androidkotlinmvvm/ui/multiple_view/SchoolListing.kt
@@ -1,4 +1,3 @@
 package com.techpaliyal.androidkotlinmvvm.ui.multiple_view
 
-interface SchoolListing {
-}
+interface SchoolListing
