+++ build.gradle
@@ -5,13 +5,17 @@ buildscript {
     repositories {
         google()
         jcenter()
-        
+        mavenCentral()
+        gradlePluginPortal()
+
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:4.2.2'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.5.30'
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
+        classpath "io.gitlab.arturbosch.detekt:detekt-gradle-plugin:1.18.1"
+
     }
 }
