+++ universal_adapter/build.gradle
@@ -2,6 +2,8 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-kapt'
+    id 'io.gitlab.arturbosch.detekt'
+
 }
 
 android {
