+++ universal_adapter/src/main/java/com/yogeshpaliyal/universalAdapter/adapter/LoadingAdapter.kt
@@ -1,4 +1,4 @@
-package com.yogeshpaliyal.universal_adapter.adapter
+package com.yogeshpaliyal.universalAdapter.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -9,7 +9,7 @@ import androidx.recyclerview.widget.AsyncDifferConfig
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.yogeshpaliyal.universal_adapter.BR
+import com.yogeshpaliyal.universalAdapter.BR
 
 
 /*
@@ -18,12 +18,11 @@ import com.yogeshpaliyal.universal_adapter.BR
 * https://techpaliyal.com
 * created on 02-05-2021 19:57
 */
-class ErrorAdapter<T>(
+class LoadingAdapter<T>(
     val lifecycleOwner: LifecycleOwner?,
-    var options : UniversalAdapterViewType.Error<T>,
-    var message: String = ""
+    var options: UniversalAdapterViewType.Loading<T>
 ) :
-    ListAdapter<T, ErrorAdapter<T>.ViewHolder>(AsyncDifferConfig.Builder<T>(object :
+    ListAdapter<T, LoadingAdapter<T>.ViewHolder>(AsyncDifferConfig.Builder<T>(object :
         DiffUtil.ItemCallback<T>() {
         override fun areItemsTheSame(oldItem: T, newItem: T): Boolean {
             return false
@@ -35,33 +34,14 @@ class ErrorAdapter<T>(
 
     }).build()) {
 
-    fun updateData(message: String?){
-        this.message = message ?: ""
-        notifyDataSetChanged()
-    }
-
-    override fun getItemViewType(position: Int): Int {
-        return options.errorLayout ?: 0
-    }
 
     inner class ViewHolder(val binding: ViewDataBinding) : RecyclerView.ViewHolder(binding.root) {
         fun bind() {
             binding.lifecycleOwner = lifecycleOwner
+            binding.setVariable(BR.binding, binding)
 
-            if (options.customBindingMapping == null) {
-                binding.setVariable(BR.message, message)
-                binding.setVariable(BR.listener, options.listener)
-                binding.setVariable(BR.binding, binding)
-                options.additionalParams?.forEach {
-                    binding.setVariable(it.key, it.value)
-                }
-                binding.executePendingBindings()
-            } else {
-
-                options.customBindingMapping?.invoke(
-                    binding,
-                    message
-                )
+            options.additionalParams?.forEach {
+                binding.setVariable(it.key, it.value)
             }
 
         }
@@ -70,16 +50,20 @@ class ErrorAdapter<T>(
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding = DataBindingUtil.inflate<ViewDataBinding>(
             LayoutInflater.from(parent.context),
-            options.errorLayout!!,
+            options.resourceLoading!!,
             parent,
             false
         )
         return ViewHolder(binding)
     }
 
+    override fun getItemViewType(position: Int): Int {
+        return options.resourceLoading ?: 0
+    }
+
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
         holder.bind()
     }
 
-    override fun getItemCount(): Int = 1
+    override fun getItemCount(): Int = options.defaultLoadingItems
 }
