+++ universal_adapter/src/main/java/com/yogeshpaliyal/universalAdapter/utils/Constans.kt
@@ -1,0 +1,4 @@
+package com.yogeshpaliyal.universalAdapter.utils
+
+
+const val DEFAULT_LOADING_ITEMS = 5
