+++ universal_adapter/src/main/java/com/yogeshpaliyal/universalAdapter/utils/Status.kt
@@ -1,4 +1,4 @@
-package com.yogeshpaliyal.universal_adapter.utils
+package com.yogeshpaliyal.universalAdapter.utils
 
 
 /**
