--- universal_adapter/src/main/java/com/yogeshpaliyal/universal_adapter/listener/TestListener.kt
@@ -1,10 +1,0 @@
-package com.yogeshpaliyal.universal_adapter.listener
-
-
-/*
-* @author Yogesh Paliyal
-* yogeshpaliyal.foss@gmail.com
-* https://techpaliyal.com
-* created on 02-12-2020 21:02
-*/
-interface TestListener
