--- universal_adapter/src/main/java/com/yogeshpaliyal/universal_adapter/listener/UniversalAdapterInterface.kt
@@ -1,4 +1,0 @@
-package com.yogeshpaliyal.universal_adapter.listener
-
-interface UniversalAdapterInterface {
-}
